reorganize concrete nodes in the module
