+++ configyaml/config/__init__.py
@@ -1,5 +1,10 @@
 from .base import AbstractNode
 from .list import ListNode
 from .dict import DictNode
-from .wildcard_dict import WildcardDictNode
 from .choice import ChoiceNode
+from .nodes import (
+    IntegerNode,
+    PositiveIntegerNode,
+    StringNode,
+    WildcardDictNode,
+)
