+++ configyaml/config/nodes.py
@@ -1,14 +1,22 @@
+from .base import AbstractNode
 from .dict import DictNode
 
 
 class WildcardDictNode(DictNode):
+    """A dictionary node where nearly any key is valid
+
+       The only invalid keys are:
+       - an actual asterisk '*'
+       - a key that begins with '_'
+       """
     def _validate_value(self):
         for key, value in self._value.items():
             # where key name doesn't matter (ex. groups)
             key_valid, explanation = self._key_name_is_valid(key)
             if key_valid:
                 field_class = self._dict_fields['*']['class']
-                field = field_class(value=value, value_node=self._find_node_for_key_value(key), context=self._context, key=key, parent=self)
+                field = field_class(value=value, value_node=self._find_node_for_key_value(key),
+                                    context=self._context, key=key, parent=self)
 
                 # don't set __dict__ if they can use any key
                 self._children[key] = field
@@ -39,3 +47,25 @@ class WildcardDictNode(DictNode):
             return False, 'Cannot start field name with a "_"'
 
         return True, 'Valid'
+
+
+class StringNode(AbstractNode):
+    """A node that must validate as a string"""
+    def __init__(self, *args, **kwargs):
+        self._type = str
+        super(StringNode, self).__init__(*args, **kwargs)
+
+
+class IntegerNode(AbstractNode):
+    """A node that must validate as an integer"""
+    def __init__(self, *args, **kwargs):
+        self._type = int
+        super(IntegerNode, self).__init__(*args, **kwargs)
+
+
+class PositiveIntegerNode(IntegerNode):
+    """A node that must validate as a positive integer"""
+    def _validate_value(self):
+        if self._value < 0:
+            description = "{value} must be a positive integer".format(value=self._value)
+            self._add_error(title="Invalid Value", description=description)
