+++ tests/test_imports.py
@@ -6,5 +6,8 @@ from configyaml.validator import ConfigValidator
 from configyaml.config import AbstractNode
 from configyaml.config import ListNode
 from configyaml.config import DictNode
-from configyaml.config import WildcardDictNode
 from configyaml.config import ChoiceNode
+from configyaml.config import IntegerNode
+from configyaml.config import PositiveIntegerNode
+from configyaml.config import StringNode
+from configyaml.config import WildcardDictNode
