+++ tests/test_list.py
@@ -1,4 +1,5 @@
 from __future__ import unicode_literals
+
 import pytest
 
 from configyaml.config import ListNode
