+++ tests/test_nodes.py
@@ -1,9 +1,23 @@
-from config_loader.config.nodes import PositiveIntegerNode, StringNode
-from config_loader.loader import ConfigLoader
+from configyaml.config import PositiveIntegerNode, StringNode, WildcardDictNode
+from configyaml.loader import ConfigLoader
 
 
 class DummyPosIntLoader(ConfigLoader):
     config_root_class = PositiveIntegerNode
+
+
+def test_wildcard_dict_invalid_key():
+    value = {'_bad': 'key'}
+    c = WildcardDictNode(value=value)
+    assert not c.is_valid()
+    assert c._errors[0].title == 'Invalid field name'
+    assert c._errors[0].description == 'Cannot start field name with a "_"'
+
+    value = {'*': 'key'}
+    c = WildcardDictNode(value=value)
+    assert not c.is_valid()
+    assert c._errors[0].title == 'Invalid field name'
+    assert c._errors[0].description == 'Field name cannot be "*"'
 
 
 def test_pos_int_invalid_validation():
